+++ NEWS.md
@@ -4,3 +4,4 @@ centerline 0.0.0.9000 (2024-03-01)
 ### NEW FEATURES
 
   * Core functionality is added through the `cent_skeleton` function
+  * Super easy routing is implemented with `cent_path` function, which can be applied in cases with predefined starting and ending points